requirements.txt
google-cloud-bigquery
prefixspan
matplotlib
pandas
umap-learn
hdbscan
scikit-learn
tqdm
pynndescent==0.5.8
bertopic
openai
InstructorEmbedding
